Rename sessionslist to sessionlist
