+++ src/components/sessions/SessionListItem.js
