+++ src/index.js
@@ -3,7 +3,7 @@ import ReactDOM from 'react-dom';
 import { Route, Switch, BrowserRouter as Router } from 'react-router-dom';
 
 import Home from './components/Home';
-import SessionsList from './components/sessions/SessionsList';
+import SessionList from './components/sessions/SessionList';
 import Session from './components/sessions/Session';
 import NewSession from './components/sessions/NewSession';
 
@@ -12,7 +12,7 @@ const routing = (
         <div>
             <Switch>
                 <Route exact path="/" component={Home}/>
-                <Route exact path="/sessions" component={SessionsList}/>
+                <Route exact path="/sessions" component={SessionList}/>
                 <Route path="/sessions/:id" component={Session}/>
                 <Route path="/new-session" component={NewSession}/>
             </Switch>
